+++ compass.rb
@@ -1,0 +1,4 @@
+environment = :development
+output_style = :nested
+sourcemap = true
+sass_options = {:precision => 10}